+++ ng-rosetta/fruit20-15/src/js/components/index.d.ts
@@ -1,6 +1,0 @@
-// This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.MD for details.
-// Copyright Thomson Reuters 2017. All rights reserved.
-
-/// <reference path="../../../typings/index.d.ts" />
-
-// Note that we moved the rest of the application-specific references out to the various files as import statements to allow webpack to identify them.