+++ ng-rosetta/fruit20-15/src/legacy.ts
@@ -1,6 +1,0 @@
-
-// In order to bootstrap Angular 1.X, we first need to bring the components down via npm in packag.json manifest.
-// Then, add the relevant typings into the typings.json manifest.
-
-// Core access point angular 1 module.
-import './js/components/app.ts';