+++ ng-rosetta/fruit20-15/src/legacy/fruit15/services/fruit-utils.ts
@@ -1,0 +1,21 @@
+// This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.
+// See the project's LICENSE.MD for details. Copyright Thomson Reuters 2017. All rights reserved.
+
+import { FruitItem } from '../index';
+
+export class FruitUtils {
+    public ExtractFruitItem(fruitList: any[], fruitItemID: number): FruitItem {
+        let fruitItem: FruitItem;
+
+        // Then, iterate through each item, looking for the specified ID.
+        fruitList.forEach((element: any) => {
+            if (element.id && element.id === fruitItemID) {
+
+                // When we find the correct item, cast it as the correct container type and return it.
+                fruitItem = new FruitItem(element);
+            }
+        });
+
+        return fruitItem;
+    }
+}