+++ ng-rosetta/fruit20-15/src/legacy/legacy-vendor.ts
@@ -1,0 +1,21 @@
+// This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.
+// See the project's LICENSE.MD for details. Copyright Thomson Reuters 2017. All rights reserved.
+
+// AngularJS References
+import 'angular';
+import 'angular-animate';
+import 'angular-sanitize';
+import 'angular-translate';
+import 'angular-translate-loader-static-files';
+import 'angular-dynamic-locale';
+
+// UI-Router
+import '@uirouter/angularjs';
+
+// Bootstrap & UI-Bootstrap
+import 'bootstrap/dist/css/bootstrap.min.css';
+import 'angular-ui-bootstrap/dist/ui-bootstrap-tpls.js';
+
+// UI Grid
+import 'angular-ui-grid/ui-grid.min.css';
+import 'angular-ui-grid/ui-grid.min.js';