+++ ng-rosetta/fruit20-15/src/legacy/legacy.ts
@@ -1,8 +1,5 @@
 // This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.
 // See the project's LICENSE.MD for details. Copyright Thomson Reuters 2017. All rights reserved.
 
-// Specify which polyfills to load from node_modules.
-
-import 'core-js/es6';
-import 'core-js/es7/reflect';
-import 'zone.js/dist/zone';
+// Core access point angular 1 module.
+import './fruit15/app.ts';