+++ ng-rosetta/fruit20-15/src/vendor.ts
@@ -1,7 +1,15 @@
-//Angular 1
-// import 'angular';
-// import 'angular-ui-router';
-// import '@angular/upgrade/static';
+// This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.
+// See the project's LICENSE.MD for details. Copyright Thomson Reuters 2017. All rights reserved.
+
+// Specify which features of Angular we will require.
+import '@angular/platform-browser';
+import '@angular/platform-browser-dynamic';
+import '@angular/core';
+import '@angular/common';
+import '@angular/upgrade/static';
 
 // Other vendors for example jQuery, Lodash or Bootstrap
 // You can import js, ts, css, sass, ...
+
+// Specify any other vendor libraries to load.
+import 'rxjs';